requirements.txt
# AstraMed Assist - Python Dependencies
# Install: pip install -r requirements.txt

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.0.0

# PDF Generation
reportlab>=4.0.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Utilities
PyYAML>=6.0
python-dotenv>=1.0.0
